LayoutTests/accessibility/checkbox-mixed-value.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="content">

<input type="checkbox" id="element1" indeterminate="true">
<input type="checkbox" id="element2" indeterminate="false">
<input type="checkbox" id="element3">

</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("Tests whether mixed values are reported properly on native checkboxes.");

    if (window.accessibilityController) {
          for (var i = 1; i <= 3; i++) {
              var element = accessibilityController.accessibleElementById("element" + i);
              debug(i + ": " + element.role);
              debug("Indeterminate status: " + element.isIndeterminate + "\n");
          }

          document.getElementById("content").style.visibility = "hidden";
    }

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
