LayoutTests/accessibility/checkbox-mixed-value-expected.txt

Tests whether mixed values are reported properly on native checkboxes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


1: AXRole: AXCheckBox
Indeterminate status: true

2: AXRole: AXCheckBox
Indeterminate status: false

3: AXRole: AXCheckBox
Indeterminate status: false

PASS successfullyParsed is true

TEST COMPLETE
